requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
influxdb==5.3.2
msgpack==1.1.0
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
urllib3==2.2.3
